Change module name to reflect Github repo URL
The previous module name was plain 'go-logstash' and this needed to be
connected to the actual Github repository URL to make this more findable
and identifiable whilst importing or using as a package in other modules.
This will also help in easier package versioning and updates under solid
namespace, enhancing the project's maintainability.